refactor: centralize config helpers
