+++ services/common/__init__.py
@@ -1,3 +1,4 @@
+from .config_utils import common_init, create_config_methods
 from .model_registry import ModelRegistry
 
-__all__ = ["ModelRegistry"]
+__all__ = ["ModelRegistry", "create_config_methods", "common_init"]
