+++ services/door_mapping_service.py
@@ -3,6 +3,8 @@ Door Mapping Service - Business logic for device attribute assignment
 Handles AI model data processing and manual override management
 """
 
+from __future__ import annotations
+
 import logging
 from dataclasses import dataclass
 from datetime import datetime
@@ -13,6 +15,7 @@ import pandas as pd
 # ADD after existing imports
 from services.ai_device_generator import AIDeviceGenerator
 from services.common import ModelRegistry
+from services.common.config_utils import common_init, create_config_methods
 from services.configuration_service import (
     ConfigurationServiceProtocol,
     DynamicConfigurationService,
@@ -20,20 +23,6 @@ from services.configuration_service import (
 from services.learning.src.api.consolidated_service import get_learning_service
 
 logger = logging.getLogger(__name__)
-
-
-def create_config_methods(cls):
-    cls.get_ai_confidence_threshold = lambda self: self.ai_threshold
-    cls.get_max_upload_size_mb = lambda self: self.max_size_mb
-    cls.get_upload_chunk_size = lambda self: self.chunk_size
-    return cls
-
-
-def common_init(self, config=None):
-    self.config = config or {}
-    self.max_size_mb = self.config.get("max_upload_size_mb", 100)
-    self.ai_threshold = self.config.get("ai_confidence_threshold", 0.8)
-    self.chunk_size = self.config.get("upload_chunk_size", 1048576)
 
 
 @dataclass
