+++ upload_validator.py
@@ -1,27 +1,14 @@
 """Compatibility wrapper around :class:`UnifiedUploadValidator`."""
+
 from __future__ import annotations
 
-from typing import Optional, Any
+from typing import Any, Optional
 
 from config.dynamic_config import dynamic_config
 from core.protocols import ConfigurationProtocol
-
+from services.common.config_utils import common_init, create_config_methods
 from services.data_processing.unified_upload_validator import UnifiedUploadValidator
 from upload_types import ValidationResult
-
-
-def create_config_methods(cls):
-    cls.get_ai_confidence_threshold = lambda self: self.ai_threshold
-    cls.get_max_upload_size_mb = lambda self: self.max_size_mb
-    cls.get_upload_chunk_size = lambda self: self.chunk_size
-    return cls
-
-
-def common_init(self, config=None):
-    self.config = config or {}
-    self.max_size_mb = self.config.get("max_upload_size_mb", 100)
-    self.ai_threshold = self.config.get("ai_confidence_threshold", 0.8)
-    self.chunk_size = self.config.get("upload_chunk_size", 1048576)
 
 UploadValidator = UnifiedUploadValidator
 
@@ -38,4 +25,3 @@ class UploadValidator(UnifiedUploadValidator):
         common_init(self, config)
         if max_size_mb is not None:
             self.max_size_mb = max_size_mb
-
